Update examples throughout document
Just updated some of the examples in the document to match our naming
conventions. Also changed the example game name from `GenericShooter` to
`MyGame`. My plan is to finish this up by fully removing all references
like weapons, etc. from the examples as it's just not that fitting with
who we are as a studio.