Finished first pass
A decent first pass on modifications to the style guide for our
purposes. Will continue to tweak further, but should serve as a good
basis now -- especially regarding structure, with the table of contents
being pretty much the first thing in the document and an extra quick
link to the asset name modifiers (arguably the most referenced section).